+++ driver-core/CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 ------
 
 - [new] The new AbsractSession makes mocking of Session easier.
+- [new] Allow to trigger a refresh of connected hosts (JAVA-309)
 
 
 2.0.2: